Disk Subsystem: Major changes to the entire subsystem
Bloated parts of the code are removed, so now the Storage Controllers
are only reading data into a buffer.
The responsibility for getting data from an offset is shifted to the
Storage Devices layer.
The Controllers layer is agnostic to how data will be
used in the upper layers (e.g. Storage Devices, Main app).

Also, while detecting disks, now the disk subsystem
is responsible for echoing information to the terminal.

In addition, now each type of Storage Device is only
overriding the hardware identification method,
so read & read_cache methods are implemented just one time.